% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_data.R
\name{annotate_data}
\alias{annotate_data}
\title{Annotate a Data Set by Combining it with a Plate Map}
\usage{
annotate_data(d, platemap, ...)
}
\arguments{
\item{d}{A data frame representing an un-annotated data set}

\item{platemap}{A data frame containing a plate map}

\item{...}{Additional arguments (not currently used)}
}
\value{
A data frame (annotated)
}
\description{
Both \code{d} and \code{platemap} should be data frames that contain columns
\code{Drug} (the 3-letter antibiotic name) and \code{Well} (e.g., "A4").
}
\examples{
\dontrun{
annotate_data <- function(my_raw_data, my_platemap)
}
}
