% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_section_spreadsheet_title.R
\name{build_section_spreadsheet_title}
\alias{build_section_spreadsheet_title}
\title{Title TODO}
\usage{
build_section_spreadsheet_title(course, year, quarter, section)
}
\arguments{
\item{course}{The course number (e.g., 180 or 200)}

\item{year}{The year}

\item{quarter}{The quarter. One of: "AU", "WI", "SP", "SU".}

\item{section}{The section name (e.g., "G")}
}
\value{
TODO
}
\description{
Title TODO
}
\examples{
build_section_title(180, 2017, "AU", "C")
}
