% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_section_data.R
\name{combine_section_data_key}
\alias{combine_section_data_key}
\alias{combine_section_data_url}
\title{Combine and Retrieve Data From Multiple Sections}
\usage{
combine_section_data_key(keys, remove_duplicates = TRUE)

combine_section_data_url(urls, remove_duplicates = TRUE)
}
\arguments{
\item{keys}{List of one or more sheet keys}

\item{remove_duplicates}{Whether or not to remove sheets that are given more
than once (default: \code{TRUE})}

\item{urls}{List of one or more sheet URLs}
}
\value{
A data frame containing data from the given sections
}
\description{
\code{combine_section_data_key} combines data from multiple
sheets specified by sheet key

\code{combine_section_data_url} combines data from multiple
sheets specified by URL
}
\examples{
\dontrun{
my_180_sections <- list_course_spreadsheets(course = 180)
all_180_data <- combine_section_data_key(my_180_sections$sheet_key)
}
}
\seealso{
\code{\link{list_course_spreadsheets}}
}
