% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_spreadsheet.R
\name{create_master_spreadsheet}
\alias{create_master_spreadsheet}
\title{Create a Master Spreadsheet (180 + 200)}
\usage{
create_master_spreadsheet(title = "Intro Bio Cure Master", trim = FALSE,
  ...)
}
\arguments{
\item{title}{Title for the spreadsheet (default: "Intro Bio Cure Master")}

\item{trim}{Should the resulting worksheet only include the necessary cells? (default: FALSE)}

\item{...}{Additional arguments passed to \code{\link[googlesheets]{gs_new}}}
}
\value{
A \code{\link[googlesheets]{googlesheet}} object
}
\description{
Create a Master Spreadsheet (180 + 200)
}
\examples{
\dontrun{
ms <- create_master_spreadsheet()
}
}
