% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_section_spreadsheet_200.R
\name{create_section_spreadsheet_200}
\alias{create_section_spreadsheet_200}
\title{Create a Google Sheet for a BIO 200 section}
\usage{
create_section_spreadsheet_200(master, year, quarter, section, num_groups,
  create_group_0 = TRUE, trim = TRUE, ...)
}
\arguments{
\item{master}{URL of the master spreadsheet}

\item{year}{The year}

\item{quarter}{The quarter. One of: "AU", "WI", "SP", "SU".}

\item{section}{The section name (e.g., "G"). Section names are named with a single letter from A-Z. When more than 26 sections exist, subsequent sections are named AA-AZ.}

\item{num_groups}{The number of groups in the section}

\item{create_group_0}{Whether or not to create a group 0 for controls (default: TRUE)}

\item{trim}{Should the resulting worksheet only include the necessary cells? (default: TRUE)}

\item{...}{Additional arguments passed to \code{\link[googlesheets]{gs_new}}}
}
\value{
A \code{\link[googlesheets]{googlesheet}} object
}
\description{
Create a Google Sheet for a BIO 200 section
}
\examples{
\dontrun{
s4 <- create_section_spreadsheet_200(year = 2017,
                                     quarter = "WI",
                                     section = "C",
                                     num_groups = 5,
                                     create_group_0 = TRUE,
                                     trim = TRUE,
                                     ...)
}
}
