% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spreadsheets_180.R
\name{create_spreadsheets_180}
\alias{create_spreadsheets_180}
\title{Create all Section Spreadsheets for BIO 180 for a Given Quarter}
\usage{
create_spreadsheets_180(master, year, quarter, num_sections, num_groups = 6,
  create_group_0 = FALSE, drugs_iso = c("None", "RIF", "STR"),
  trim = TRUE)
}
\arguments{
\item{master}{URL of the master spreadsheet}

\item{year}{The year}

\item{quarter}{The quarter. One of: "AU", "WI", "SP", "SU".}

\item{num_sections}{Total number of sections}

\item{num_groups}{Number of groups for each section (default: 6)}

\item{create_group_0}{Whether or not to create a group 0 for controls (default: FALSE)}

\item{drugs_iso}{List of drugs used when creating progenitors (default: \code{c("None", "RIF", "STR")})}

\item{trim}{Should the resulting worksheet only include the necessary cells? (default: TRUE)}
}
\value{
A list of \code{\link[googlesheets]{googlesheet}} objects, one per section
}
\description{
Create all Section Spreadsheets for BIO 180 for a Given Quarter
}
\examples{
\dontrun{
my_180_sheets <- create_spreadsheets_180(
    master = "https://docs.google.com/spreadsheets/d/1aOG8u-r1pObn2_OiiUx530eo2iACrU4nbvQ_9gmdcUw/",
    year = 2018,
    quarter = "AU",
    num_sections = 3
)
}
}
