% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_section_spreadsheet.R
\name{delete_section_spreadsheet}
\alias{delete_section_spreadsheet}
\title{Title}
\usage{
delete_section_spreadsheet(course, year, quarter, section, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
