% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_master_data.R
\name{get_master_data}
\alias{get_master_data}
\alias{save_master_data}
\title{Retrieve Data from Master Spreadsheet}
\usage{
get_master_data(master, problems_as_logical = FALSE, ...)

save_master_data(master, path, ...)
}
\arguments{
\item{master}{URL of the master spreadsheet}

\item{problems_as_logical}{Whether or not to try to convert ProblemIdentified
and SequenceProblemIdentified values to logical (default: \code{FALSE})}

\item{...}{Additional arguments passed to \code{\link[googlesheets]{gs_read}} (\code{get_master_data}) or \code{\link[readr]{write_csv}} (\code{save_master_data}).}

\item{path}{File path or connection to write data to}
}
\value{
A data frame (tibble)
}
\description{
\code{get_master_data} returns the data from the master spreadsheet as a data
frame ()

\code{save_master_data} saves data from the master spreadsheet
to a csv file
}
\examples{
\dontrun{
TODO
}
}
