% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next_strainid.R
\name{get_next_strainid}
\alias{get_next_strainid}
\title{Title}
\usage{
get_next_strainid(mastersheet_title = "Intro Bio Cure Master",
  numeric = FALSE)
}
\arguments{
\item{mastersheet_title}{Title of the master spreadsheet (default: "Intro Bio Cure Master")}

\item{numeric}{Whether or not the result should be numeric (default:
\code{FALSE}). If \code{FALSE}, a strain ID is returned.}
}
\value{
An integer specifying the number of the next available strain ID
}
\description{
Title
}
\examples{
next_strain <- get_next_strainid()
next_strain_id <- number_as_strainid(next_strain)

}
