% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_section_data.R
\name{get_section_data}
\alias{get_section_data}
\title{TODO}
\usage{
get_section_data(course, year, quarter, section, include_course = TRUE, ...)
}
\arguments{
\item{course}{The course number (e.g., 180 or 200)}

\item{year}{The year}

\item{quarter}{The quarter. One of: "AU", "WI", "SP", "SU".}

\item{section}{The section name (e.g., "G"). Section names are named with a single letter from A-Z. When more than 26 sections exist, subsequent sections are named AA-AZ.}

\item{include_course}{Whether or not to include the course number as a column (default: TRUE)}

\item{...}{Additional arguments passed to \code{\link[googlesheets]{gs_read}}}
}
\value{
TODO
}
\description{
TODO
}
\examples{
#TODO
}
