% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_section_data_180.R
\name{get_section_data_180}
\alias{get_section_data_180}
\alias{get_section_data_180_key}
\title{Get Data for a Given Section}
\usage{
get_section_data_180(url, remove_duplicates = TRUE, ...)

get_section_data_180_key(key, remove_duplicates = TRUE, ...)
}
\arguments{
\item{url}{One or more sheet URLs}

\item{remove_duplicates}{Whether or not to remove sheets that are given more
than once (default: \code{TRUE})}

\item{...}{Additional arguments (not currently used)}

\item{key}{One or more sheet keys}
}
\value{
A data frame
}
\description{
Both \code{get_section_data_180} and \code{get_section_data_180_key} retrieve
data for a given section (or sections). \code{get_section_data_180} is given
the URL(s) of section sheets, while \code{get_section_data_180_key} is given
the unique keys.
}
\examples{
\dontrun{
# TODO
}
}
