% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_platemaps.R
\name{load_platemaps}
\alias{load_platemaps}
\title{Load One or More Plate Maps}
\usage{
load_platemaps(..., include_drug = TRUE)
}
\arguments{
\item{...}{One or more pairs of drug abbreviation and file name. See example.}

\item{include_drug}{Whether or not to include the drug name as a column (default: \code{TRUE}).}
}
\value{
A list of data frames, one for each drug.
}
\description{
Plate map files should be CSVs that contain at least the following columns: Well, Pro.or.Des, Drug.at.Isolation, and Concentration.
}
\examples{
\dontrun{
maps <- load_platemaps(STR = "my_strep_pmap.csv", RIF = "my_rif_pmap.csv")
}
}
