% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_MIC_control_plots.R
\name{make_MIC_control_plots}
\alias{make_MIC_control_plots}
\title{Create MIC Plate Diagrams for Control Plates}
\usage{
make_MIC_control_plots(control_data, path = ".", absorbance_max = 3, ...)
}
\arguments{
\item{control_data}{Data frame containing absorbance values for control plates}

\item{path}{Path in which to place plots (default: \code{"."}, the current working directory)}

\item{absorbance_max}{Value to be set as the maximum possible absorbance. Useful for having consistent greyscale across many plots (default: \code{3}, which is approx. where the spex maxes out).}

\item{...}{Additional arguments, passed to \code{\link{plot_mic_plate}}}
}
\description{
Create MIC Plate Diagrams for Control Plates
}
\examples{
\dontrun{
make_MIC_control_plots(my_control_data)
}
}
