% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_MIC_section_plots.R
\name{make_MIC_section_plots}
\alias{make_MIC_section_plots}
\title{Create MIC Plate Diagrams for Class Data}
\usage{
make_MIC_section_plots(plate_data, path = ".", absorbance_max = 3,
  make_section_dirs = TRUE)
}
\arguments{
\item{plate_data}{Data frame containing MIC assay data}

\item{path}{Path in which to place plots (default: \code{"."}, the current working directory)}

\item{absorbance_max}{Value to be set as the maximum possible absorbance. Useful for having consistent greyscale across many plots (default: \code{3}, which is approx. where the spex maxes out).}

\item{make_section_dirs}{Whether or not to create subdirectories for each section (default: \code{TRUE})}
}
\description{
Create MIC Plate Diagrams for Class Data
}
\examples{
\dontrun{
make_MIC_section_plots(my_plate_data)
}
}
