% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_add_section_data_180.R
\name{master_add_section_data_180}
\alias{master_add_section_data_180}
\title{Insert Section Data into Master Sheet}
\usage{
master_add_section_data_180(title, d)
}
\arguments{
\item{title}{Title of the master spreadsheet (default: "Intro Bio Cure Master")}

\item{d}{Data frame (see \code{get_section_data})}
}
\value{
TODO
}
\description{
Insert Section Data into Master Sheet
}
\examples{
\dontrun{TODO}
}
