% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_add_section_data_200.R
\name{master_add_section_data_200}
\alias{master_add_section_data_200}
\title{Insert Section Data from BIO 200 into Master Sheet}
\usage{
master_add_section_data_200(master, d)
}
\arguments{
\item{master}{URL of the master spreadsheet}

\item{d}{Data frame containing section data (see \code{\link{get_section_data_180}})}
}
\description{
Insert Section Data from BIO 200 into Master Sheet
}
\examples{
\dontrun{
d <- get_section_data_200(<section url>)
master_add_section_data_200(<master sheet url>, d)
}
}
