% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sheet_title.R
\name{parse_sheet_title}
\alias{parse_sheet_title}
\title{Parse a Course Spreadsheet Title}
\usage{
parse_sheet_title(title)
}
\arguments{
\item{title}{Title for a course spreadsheet (e.g., "BIO200 WI2018 Section B")}
}
\value{
A list of lists, where each list contains information about the
corresponding title
}
\description{
Parse a Course Spreadsheet Title
}
\examples{
parse_sheet_title("BIO200 WI2018 Section B")

}
