% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mic_plate.R
\name{plot_mic_plate}
\alias{plot_mic_plate}
\title{Make a MIC Plate Diagram}
\usage{
plot_mic_plate(d, title = NULL, subtitle = NULL, threshold = 0,
  absorbance.max = 3, show.values = FALSE, show.unused = FALSE, ...)
}
\arguments{
\item{d}{A data frame containing absorbance and antibiotic data}

\item{title}{A title for the plot. If not provided, one will be created using the \code{Section} and \code{Group}.}

\item{subtitle}{A subtitle for the plot. If not provided, one will be created using the \code{Drug}.}

\item{threshold}{Value above which values can be trusted. If a value is below the threshold, that well will be highlighted in red (default: \code{0}).}

\item{absorbance.max}{Value to be set as the maximum possible absorbance. Useful for having consistent greyscale across many plots (default: \code{3}, which is approx. where the spex maxes out).}

\item{show.values}{Whether or not to display the absorbance values for each well (default: \code{FALSE})}

\item{show.unused}{Whether or not to display unused wells. (default: \code{FALSE})}

\item{...}{Additional arguments (not currently used)}
}
\value{
A \code{\link{ggplot2}} plot object
}
\description{
\code{plot_mic_plate} makes a MIC Plate Diagram using the data supplied. This
data frame must contain the columns described below.
}
\details{
The following columns must be present in the data:

\describe{
    \item{Drug}{3-letter code indicating the drug used for the MIC assay (e.g., 'RIF' or 'STR'). Must have one unique value.}
    \item{Well}{Well in the microtiter plate (e.g., 'D6')}
    \item{Pro.or.Des}{Strain information. Either 'Progenitor' or 'Descendant'}
    \item{Drug.at.Isolation}{Drug used when strain was isolated, either 'None' or 3-letter code (e.g., 'RIF')}
    \item{Absorbance}{Absorbance value for the well}
    \item{Concentration}{Concentration of antibiotic present in the well}
}

Additional columns that can be used:

\describe{
    \item{Section}{BIO 180 Section name (e.g., 'A'). Must have one unique value.}
    \item{Group}{BIO 180 Group number. Must have one unique value.}
    \item{Strain}{Strain name (e.g., 'RIF.R.D'). If not present, this will be created using \code{Pro.or.Des} and \code{Drug.at.Isolation}}
}
}
\examples{
\dontrun{
plot_mic_plate(my_mic_data)
}
}
