% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spec_data.R
\name{read_spec_data}
\alias{read_spec_data}
\title{Read Data from XML Files Exported by Spectrophotometer}
\usage{
read_spec_data(file, ...)
}
\arguments{
\item{file}{The name of the file to read}

\item{...}{Additional arguments (not currently used)}
}
\value{
A data frame containing all data from the reads. Column
\code{Absorbance.Adj} will be the absorbance data from a culture minus the
absorbance of that same well in the control plate (of that drug). If no
control plate exists for that drug, then \code{NA} is given.
}
\description{
Read Data from XML Files Exported by Spectrophotometer
}
\examples{
\dontrun{
dSpec <- read_spec_data(file = "myfile.xml")
}
}
