% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{gs_auth}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{googlesheets}{\code{\link[googlesheets]{gs_auth}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

