% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_as_number.R
\name{section_as_number}
\alias{section_as_number}
\alias{number_as_section}
\title{Convert Numbers to/from Section Names}
\usage{
section_as_number(x)

number_as_section(x)
}
\arguments{
\item{x}{A number (\code{number_as_section}) or strain ID (\code{section_as_number})}
}
\value{
\code{number_as_ssection} returns a string, and \code{section_as_number} returns a numeric value
}
\description{
Convert Numbers to/from Section Names
}
\examples{
section_as_number("C")
number_as_section(36)

}
