% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_station_sheets.R
\name{select_section_sheets}
\alias{select_section_sheets}
\title{Select Section Sheets Based on Course, Year, Quarter, and/or Section}
\usage{
select_section_sheets(course = NULL, year = NULL, quarter = NULL,
  section = NULL)
}
\arguments{
\item{course}{The course number (e.g., 180 or 200) (optional)}

\item{year}{The year (optional)}

\item{quarter}{The quarter (optional)}

\item{section}{The section name (optional)}
}
\value{
A vector of spreadsheet keys
}
\description{
\code{select_section_sheets} selects course spreadsheets. If any arguments
are provided, only sheets matching those parameters are shown. For example,
if \code{course = 180}, only sheets for BIO 180 are returned
}
\examples{
\dontrun{
select_section_sheets(course = 180)
}
}
