% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_name.R
\name{strain_name}
\alias{strain_name}
\title{Create a Strain Name Based on Resistance Status and Time}
\usage{
strain_name(Pro.or.Des, Drug.at.Isolation)
}
\arguments{
\item{Pro.or.Des}{Character, either \code{"Progenitor"} or \code{"Descendant"}}

\item{Drug.at.Isolation}{Character, 3-character drug name (e.g., \code{"RIF"})}
}
\value{
A string
}
\description{
\code{strain_name} creates a strain name based on \code{Pro.or.Des} and
\code{Drug.at.Isolation} values. For example, (\code{"Progenitor"},
\code{"RIF"}) becomes \code{"RIF.R"}, while (\code{"Descendant"},
\code{"None"}) becomes \code{"SEN.D"}
}
\examples{
strain_name("Progenitor", "RIF")
}
