% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainid_as_number.R
\name{strainid_as_number}
\alias{strainid_as_number}
\alias{number_as_strainid}
\title{Convert Numbers to/from Strain IDs}
\usage{
strainid_as_number(x)

number_as_strainid(x)
}
\arguments{
\item{x}{A number (\code{number_as_strainid}) or strain ID (\code{strainid_as_number})}
}
\value{
\code{number_as_strainid} returns a string, and \code{strainid_as_number} returns a numeric value
}
\description{
Strain IDs are 5-character strings with two uppercase letters followed by
three digits.
}
\examples{
strainid_as_number("BC828")
number_as_strainid(90210)
}
