% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_course}
\alias{is_course}
\alias{is_quarter}
\alias{is_section}
\title{Validate Values Given}
\usage{
is_course(x)

is_quarter(x)

is_section(x)
}
\arguments{
\item{x}{Value to test}
}
\value{
Logical value indicating whether or not input \code{x} is a valid
value
}
\description{
\code{is_course} determines whether or not a given course number
is valid. Valid course numbers are 180 and 200.

\code{is_quarter} determines whether or not a given quarter
name is valid. Valid values include "AU" (Autumn), "WI" (Winter),
"SP" (Spring), and "SU" (Summer)

\code{is_section} Determines whether or not a given section name
is valid. Section names are letters A-Z. When more than 26 sections exist,
two letter codes are used.
}
\examples{
is_course(180)
is_quarter("WI")
is_section("C")
}
