% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_section_spreadsheet.R
\name{view_section_spreadsheet}
\alias{view_section_spreadsheet}
\alias{view_section_spreadsheet_180}
\title{Open the Google Sheet for a BIO 180 section}
\usage{
view_section_spreadsheet(course, year, quarter, section)

view_section_spreadsheet_180(year, quarter, section)
}
\arguments{
\item{section}{}
}
\description{
Open the Google Sheet for a BIO 180 section
}
