% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_section_data.R
\name{combine_section_data_key}
\alias{combine_section_data_key}
\alias{combine_section_data_url}
\title{Combine and Retrieve Data From Multiple Sections}
\usage{
combine_section_data_key(..., remove_duplicates = TRUE)

combine_section_data_url(..., remove_duplicates = TRUE)
}
\arguments{
\item{...}{List or vector specifying one or more sheets.}

\item{remove_duplicates}{Whether or not to remove sheets that are given more
than once (default: \code{TRUE})}
}
\value{
TODO
}
\description{
\code{combine_section_data_key} combines data from multiple
sheets specified by sheet key

\code{combine_section_data_url} combines data from multiple
sheets specified by URL
}
\examples{
\dontrun{
TODO
}
}
\seealso{
\code{\link{list_course_spreadsheets}}
}
