% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_spreadsheet.R
\name{create_master_spreadsheet}
\alias{create_master_spreadsheet}
\title{Create a Master Spreadsheet (180 + 200)}
\usage{
create_master_spreadsheet(title, data = NULL, trim = FALSE, ...)
}
\arguments{
\item{title}{Title for the spreadsheet}

\item{data}{Optional, A data frame containing data to be used (default: NULL)}

\item{trim}{Should the resulting worksheet only include the necessary cells? (default: FALSE)}

\item{...}{Additional arguments passed to \code{\link[googlesheets]{gs_new}}}
}
\value{
A \code{\link[googlesheets]{googlesheet}} object
}
\description{
Create a Master Spreadsheet (180 + 200)
}
\note{
An invalid first row is also created, because this is required by googlesheets.
}
\examples{
\dontrun{
ms <- create_master_spreadsheet()
}
}
