% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_section_spreadsheet_180.R
\name{create_section_spreadsheet_180}
\alias{create_section_spreadsheet_180}
\title{Create a Google Sheet for a BIO 180 section}
\usage{
create_section_spreadsheet_180(year, quarter, section, num_groups,
  create_group_0 = FALSE, start_strainid = NULL, drugs_iso = c("None",
  "RIF", "STR"), trim = TRUE, ...)
}
\arguments{
\item{year}{The year}

\item{quarter}{The quarter. One of: "AU", "WI", "SP", "SU".}

\item{section}{The section name (e.g., "G"). Section names are named with a single letter from A-Z. When more than 26 sections exist, subsequent sections are named AA-AZ.}

\item{num_groups}{The number of groups in the section}

\item{create_group_0}{Whether or not to create a group 0 for controls (default: FALSE)}

\item{start_strainid}{First strain ID to use. Can be numeric (e.g., 1000) or a strain ID ("AB000"). If none is provided, the StrainID column will be left empty.}

\item{drugs_iso}{List of drugs used when creating progenitors (default: \code{c("None", "RIF", "STR")})}

\item{trim}{Should the resulting worksheet only include the necessary cells? (default: TRUE)}

\item{...}{Additional arguments passed to \code{\link[googlesheets]{gs_new}}}
}
\value{
A \code{\link[googlesheets]{googlesheet}} object
}
\description{
Create a Google Sheet for a BIO 180 section
}
\examples{
\dontrun{
s4 <- create_section_spreadsheet_180(year = 2017,
                                     quarter = "WI",
                                     section = "C",
                                     num_groups = 5,
                                     create_group_0 = FALSE,
                                     start_strainid = 300,
                                     trim = TRUE,
                                     ...)
}
}
