% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spreadsheets_180.R
\name{create_spreadsheets_180}
\alias{create_spreadsheets_180}
\title{Create all Section Spreadsheets for BIO 180 for a Given Quarter}
\usage{
create_spreadsheets_180(master, year, quarter, num_sections, num_groups = 6,
  create_group_0 = TRUE, drugs_iso = c("None", "Rifampicin",
  "Streptomycin"), trim = TRUE)
}
\arguments{
\item{master}{URL of the master spreadsheet}

\item{year}{The year}

\item{quarter}{The quarter. One of: "AU", "WI", "SP", "SU".}

\item{num_sections}{Total number of sections}

\item{num_groups}{Number of groups for each section (default: 6)}

\item{create_group_0}{Whether or not to create a group 0 for controls (default: TRUE)}

\item{drugs_iso}{List of drugs used when creating progenitors (default: \code{c("None", "Rifampicin", "Streptomycin")})}

\item{trim}{Should the resulting worksheet only include the necessary cells? (default: TRUE)}
}
\description{
Create all Section Spreadsheets for BIO 180 for a Given Quarter
}
\examples{
\dontrun{
create_spreadsheets_180("Intro Bio Cure Master", year = 2018, quarter = "AU", num_sections = 3)
}
}
