% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_course_spreadsheets.R
\name{list_course_spreadsheets}
\alias{list_course_spreadsheets}
\title{List Existing Course Spreadsheets}
\usage{
list_course_spreadsheets(course = NULL, year = NULL, quarter = NULL,
  section = NULL)
}
\arguments{
\item{course}{The course number (e.g., 180 or 200) (optional)}

\item{year}{The year (optional)}

\item{quarter}{The quarter (optional)}

\item{section}{The section name (optional)}
}
\value{
A data frame with one row per matching sheet
}
\description{
\code{list_course_spreadsheets} creates a table with information about
existing course spreadsheets. If any arguments are provided, only sheets
matching those parameters are shown. For example, if \code{course = 180},
only sheets for BIO 180 are returned.
}
\examples{
sheets_2017 <- list_course_spreadsheets(year = 2017)

}
