% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_add_section_data_180.R
\name{master_add_section_data_180}
\alias{master_add_section_data_180}
\title{Insert Section Data into Master Sheet}
\usage{
master_add_section_data_180(master, d)
}
\arguments{
\item{master}{URL of the master spreadsheet}

\item{d}{Data frame containing section data (see \code{\link{get_section_data_180}})}
}
\description{
Insert Section Data into Master Sheet
}
\examples{
\dontrun{
d <- get_section_data_180(<section url>)
master_add_section_data_180(<master sheet url>, d)
}
}
