% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_section_spreadsheet.R
\name{view_section_spreadsheet}
\alias{view_section_spreadsheet}
\alias{view_section_spreadsheet_180}
\alias{view_section_spreadsheet_200}
\title{Open the Google Sheet for a BIO 180/200 section}
\usage{
view_section_spreadsheet(course, year, quarter, section)

view_section_spreadsheet_180(year, quarter, section)

view_section_spreadsheet_200(year, quarter, section)
}
\arguments{
\item{course}{The course number (e.g., 180 or 200)}

\item{year}{The year}

\item{quarter}{The quarter. One of: "AU", "WI", "SP", "SU".}

\item{section}{The section name (e.g., "G"). Section names are named with a single letter from A-Z. When more than 26 sections exist, subsequent sections are named AA-AZ.}
}
\description{
Open the Google Sheet for a BIO 180/200 section
}
\examples{
\dontrun{
view_section_spreadsheet(course = 180, year = 2017, quarter = "AU", section = "C")
}
}
